backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
numpy==2.1.3
opencv-python==4.10.0.84
Pillow==11.0.0
pytest==8.3.3
sentencepiece==0.2.0
torch==2.5.1
transformers==4.46.3
